+++ src/onegov.town/onegov/town/forms/person.py
@@ -9,7 +9,7 @@ from wtforms.widgets import TextArea
 class PersonForm(Form):
     """ Form to edit people. """
 
-    academic_title = StringField(_("Academic Title"))
+    salutation = StringField(_("Salutation"))
 
     first_name = StringField(_("First name"), [validators.InputRequired()])
     last_name = StringField(_("Last name"), [validators.InputRequired()])